pyproject.toml
[project]
name = "sbifitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tharvey303", email = "tharvey303@gmail.com" }
]
requires-python = ">=3.10"

dependencies = [
    "torch",
    "lampe",
    "ltu-ili @ git+https://github.com/maho3/ltu-ili#egg=ltu-ili",
    "tarp @ git+https://github.com/maho3/tarp.git",
    "cosmos-synthesizer",
    "scikit-learn",
    "sbi>=0.22.0",
    "h5py",
    "optuna",
    "tqdm",
    "astropy",
    "numpy",
    "scipy",
    "unyt"
]

[project.optional-dependencies]

simformer = [
    "jax",
]

# Development
dev = [
    "ruff==0.11.7",
    "pre-commit",
    "nb-clean",
]
# Testing
test = [
    "nbmake",
    "pytest",
    "ruff==0.11.7",
    "mypy>=0.910",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true


[tool.hatch.build.targets.wheel]
packages=["src/sbifitter"]

# Configure the linter and formatter
[tool.ruff]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    ".DS_Store",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".pre-commit-config.yaml",
    "*__init__.py"  # ignore all init files
]

# PEP-8 line length
line-length = 90
indent-width = 4



# Assume Python 3.8 by default regardless
target-version = "py38"


[tool.ruff.lint]
select = ["F", "E", "W", "D"]
ignore = [
       "E402",  # "module level import not at top of file" (isolate C imports in case python alternatives exist)
       "F811",  # "redefinition of unused name from line N" (breaks quantity objects)
       ]


# Sort imports alphabetically
extend-select = ["I"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
# For every .py under examples/, ignore all D-rules
"examples/**/*.py" = ["D"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = 20

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

